requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
base58==2.1.1
beautifulsoup4==4.12.2
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
dataclasses-json==0.6.3
distlib==0.3.7
distro==1.8.0
ecdsa==0.18.0
faiss-cpu==1.7.4
filelock==3.13.1
firebase-admin==6.3.0
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
google-api-core==2.15.0
google-api-python-client==2.109.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
httpcore==1.0.1
httplib2==0.22.0
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.347
langchain-core==0.0.11
langsmith==0.0.69
MarkupSafe==2.1.3
marshmallow==3.20.1
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.1.1
packaging==23.2
platformdirs==4.0.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
pyparsing==3.1.1
PySimpleGUI==4.60.5
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.7
Werkzeug==3.0.1
yarl==1.9.4
